+++ PSFramework/internal/loggingProviders/sql.provider.ps1
@@ -26,43 +26,44 @@
 			$ObjectToProcess
 		)
 		
-		begin {
-			$SqlServer = Get-ConfigValue -Name 'SqlServer'
-			$SqlTable = Get-ConfigValue -Name 'Table'
-			$SqlDatabaseName = Get-ConfigValue -Name 'Database'
-			$SqlSchema = Get-ConfigValue -Name 'Schema'
-			if (-not $SqlSchema) { $SqlSchema = 'dbo' }
-		}
-		
 		process {
-			$QueryParameters = @{
-				"Message" = $ObjectToProcess.LogMessage
-				"Level"   = $ObjectToProcess.Level -as [string]
-				"TimeStamp" = $ObjectToProcess.TimeStamp.ToUniversalTime()
-				"FunctionName" = $ObjectToProcess.FunctionName
-				"ModuleName" = $ObjectToProcess.ModuleName
-				"Tags"    = $ObjectToProcess.Tags -join "," -as [string]
-				"Runspace" = $ObjectToProcess.Runspace -as [string]
-				"ComputerName" = $ObjectToProcess.ComputerName
-				"TargetObject" = $ObjectToProcess.TargetObject -as [string]
-				"File"    = $ObjectToProcess.File
-				"Line"    = $ObjectToProcess.Line
-				"ErrorRecord" = $ObjectToProcess.ErrorRecord -as [string]
-				"CallStack" = $ObjectToProcess.CallStack -as [string]
-			}
+			$queryParameters = $script:converter.Process($ObjectToProcess)
+			$insertQuery = Get-Query -Parameters $queryParameters
 			
 			try {
-				$SqlInstance = Connect-DbaInstance -SqlInstance $SqlServer
+				$SqlInstance = Connect-DbaInstance -SqlInstance $script:cfgServer
 				if ($SqlInstance.ConnectionContext.IsOpen -ne 'True') {
 					$SqlInstance.ConnectionContext.Connect() # Try to connect to the database
 				}
 				
-				$insertQuery = "INSERT INTO [$SqlDatabaseName].[$SqlSchema].[$SqlTable](Message, Level, TimeStamp, FunctionName, ModuleName, Tags, Runspace, ComputerName, TargetObject, [File], Line, ErrorRecord, CallStack)
-                VALUES (@Message, @Level, @TimeStamp, @FunctionName, @ModuleName, @Tags, @Runspace, @ComputerName, @TargetObject, @File, @Line, @ErrorRecord, @CallStack)"
-				Invoke-DbaQuery -SqlInstance $SqlInstance -Database $SqlDatabaseName -Query $insertQuery -SqlParameters $QueryParameters -EnableException
+				Invoke-DbaQuery -SqlInstance $SqlInstance -Database $script:cfgDatabase -Query $insertQuery -SqlParameters $queryParameters -EnableException
 			}
 			catch { throw }
 		}
+	}
+	
+	function Get-Query {
+		[CmdletBinding()]
+		param (
+			[hashtable]
+			$Parameters
+		)
+		
+		if ($script:insertQuery) { return $script:insertQuery }
+		
+		$properties = $Parameters.Keys
+		$propSquared = foreach ($property in $properties) {
+			"[$property]"
+		}
+		$propAdd = foreach ($property in $properties) {
+			"@$property"
+		}
+		
+		$script:insertQuery = @"
+INSERT INTO [$script:cfgDatabase].[$script:cfgSchema].[$script:cfgTable]($($propSquared -join ','))
+VALUES ($($propAdd -join ','))
+"@
+		$script:insertQuery
 	}
 	
 	function New-DefaultSqlDatabaseAndTable {
@@ -155,6 +156,67 @@ $isInstalled_script = {
 $begin_event = {
 	New-DefaultSqlDatabaseAndTable
 }
+$start_event = {
+	$changePending = $false
+	if ($script:cfgHeaders -ne (Get-ConfigValue -Name 'Headers')) {
+		$script:cfgHeaders = Get-ConfigValue -Name 'Headers'
+		$changePending = $true
+	}
+	if ($script:cfgServer -ne (Get-ConfigValue -Name 'SqlServer')) {
+		$script:cfgServer = Get-ConfigValue -Name 'SqlServer'
+		$changePending = $true
+	}
+	if ($script:cfgDatabase -ne (Get-ConfigValue -Name 'Database')) {
+		$script:cfgDatabase = Get-ConfigValue -Name 'Database'
+		$changePending = $true
+	}
+	if ($script:cfgSchema -ne (Get-ConfigValue -Name 'Schema')) {
+		$script:cfgSchema = Get-ConfigValue -Name 'Schema'
+		if (-not $script:cfgSchema) { $script:cfgSchema = 'dbo' }
+		$changePending = $true
+	}
+	if ($script:cfgTable -ne (Get-ConfigValue -Name 'Table')) {
+		$script:cfgTable = Get-ConfigValue -Name 'Table'
+		$changePending = $true
+	}
+	if (-not $changePending) { return }
+	
+	$script:sql_headers = switch ($script:cfgHeaders) {
+		'Tags'
+		{
+			@{
+				Name	   = 'Tags'
+				Expression = { ($_.Tags -join ",") -as [string] }
+			}
+		}
+		'Message' { @{ Name = 'Message'; Expression = { $_.LogMessage } } }
+		'Level' { @{ Name = 'Level'; Expression = { $_.Level -as [string] } } }
+		'Runspace' { @{ Name = 'Runspace'; Expression = { $_.Runspace -as [string] } } }
+		'TargetObject' { @{ Name = 'TargetObject'; Expression = { $_.TargetObject -as [string] } } }
+		'ErrorRecord' { @{ Name = 'ErrorRecord'; Expression = { $_.ErrorRecord -as [string] } } }
+		'CallStack' { @{ Name = 'CallStack'; Expression = { $_.CallStack -as [string] } } }
+		'Timestamp'
+		{
+			@{
+				Name						   = 'Timestamp'
+				Expression					   = {
+					$_.Timestamp.ToUniversalTime()
+				}
+			}
+		}
+		default { $_ }
+	}
+	
+	if ($script:converter) {
+		$null = $script:converter.End()
+		$script:converter = $null
+	}
+	# Cache the conversion logic once as a steppable pipeline to avoid having to do it
+	$script:converter = { Microsoft.PowerShell.Utility\Select-Object $script:sql_headers | PSFramework\ConvertTo-PSFHashtable }.GetSteppablePipeline()
+	$script:converter.Begin($true)
+	
+	$script:insertQuery = ''
+}
 
 $message_event = {
 	param (
@@ -162,6 +224,13 @@ $message_event = {
 	)
 	
 	Export-DataToSql -ObjectToProcess $Message
+}
+
+$end_event = {
+	if ($script:converter) {
+		$null = $script:converter.End()
+		$script:converter = $null
+	}
 }
 
 # Action that is performed when stopping the logging script.
@@ -185,9 +254,11 @@ $paramRegisterPSFSqlProvider = @{
 	Name			   = "Sql"
 	Version2		   = $true
 	ConfigurationRoot  = 'PSFramework.Logging.Sql'
-	InstanceProperties = 'Database', 'Schema', 'Table', 'SqlServer', 'Credential'
+	InstanceProperties = 'Database', 'Schema', 'Table', 'SqlServer', 'Credential', 'Headers'
 	MessageEvent	   = $message_Event
 	BeginEvent		   = $begin_event
+	StartEvent		   = $start_event
+	EndEvent		   = $end_event
 	FinalEvent		   = $final_event
 	IsInstalledScript  = $isInstalled_script
 	InstallationScript = $installation_script
@@ -198,6 +269,7 @@ $paramRegisterPSFSqlProvider = @{
 		'Database' = "LoggingDatabase"
 		'Table'    = "LoggingTable"
 		'Schema'   = 'dbo'
+		Headers    = 'Message', 'Timestamp', 'Level', 'Tags', 'Data', 'ComputerName', 'Runspace', 'UserName', 'ModuleName', 'FunctionName', 'File', 'CallStack', 'TargetObject', 'ErrorRecord'
 	}
 }