+++ PSFramework/internal/tepp/scripts/temp.ps1
@@ -1,0 +1,3 @@
+﻿Register-PSFTeppScriptblock -Name 'PSFramework.Temp.ProviderName' -ScriptBlock {
+	$script:tempItems.Providers.Keys
+}