+++ PSFramework/tests/general/FileIntegrity.Exceptions.ps1
@@ -21,7 +21,7 @@ $global:MayContainCommand = @{
 	"Write-Verbose" = @()
 	"Write-Warning" = @('assembly.ps1')
 	"Write-Error"  = @('Invoke-PSFCommand.ps1','Stop-PSFFunction.ps1')
-	"Write-Output" = @('filesystem.provider.ps1', 'gelf.provider.ps1', 'logfile.provider.ps1', 'input.ps1', 'teppCoreCode.ps1')
+	"Write-Output" = @('filesystem.provider.ps1', 'gelf.provider.ps1', 'Import-PSFPowerShellDataFile.ps1', 'logfile.provider.ps1', 'input.ps1', 'teppCoreCode.ps1')
 	"Write-Information" = @()
 	"Write-Debug" = @()
 }