+++ PSFramework/changelog.md
@@ -1,5 +1,20 @@
 ﻿# CHANGELOG
 
+## 1.7.226 (2022-05-13)
+
+- New: Component: Temp Item - simplifies creation and lifecycle management of temporary items - files, folders, ...
+- New: Configuration Validation: Secret - accepts either a PSCredential, a SecureString or a string, stores the secret as a PSCredential (#508)
+- New: ScriptBlock: PSFramework.Validate.SafeName - ensures input only consists of letters, numbers, dots, underscores and dashes
+- Upd: Command: Import-PSFPowerShellDataFile - added support for json files
+- Upd: Command: Import-PSFPowerShellDataFile - added `-Unsafe` parameter to allow loading psd1 files with multiple hashtables
+- Upd: Logging Provider: Azure Log Analytics - added TimeFormat option to define, just how exactly timestamps are being logged
+- Upd: Logging Provider: Azure Log Analytics - added Headers option to define, which properties (in which order) are being logged (#507)
+- Upd: Logging Provider: Azure Log Analytics - WorkspaceID & SharedKey now updated to use the secret validation, supporting SecureString or Credential objects in addition to plaintext string (#508)
+- Upd: Type: UtilityHost - added public static method: GetPrivateStaticField
+- Upd: Type: utilityHost - added public static method: GetCallerInfo to get information on whoever called your command
+- Upd: Removal-Event has been hidden from Get-Job
+- Fix: Memory leak that would affect long-running processes that utilize runspaces
+
 ## 1.6.214 (2021-11-11)
 
 - New: Command: New-PSFThrottle - Create a throttle object, used to not exceed a certain rate of executions per time interval.